% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{extractStd}
\alias{extractStd}
\title{Extract Standards}
\usage{
extractStd(MFI, stdstr, bgstr, dilut, smpdil, antigen, yvar)
}
\arguments{
\item{MFI}{data frame with variable "Sample" that contains information about
standard concentrations in "1/dilution" format (e.g. "1/200").}

\item{stdstr}{character string indicating standards in the file's "Sample"
column. Not case sensitive. If \code{""} (empty string), standards will be
determined by the pattern "1/" only.}

\item{bgstr}{character string indicating background in the file's "Sample"
column. Not case sensitive.}

\item{dilut}{standard dilutions to use for standard curve fitting. If
\code{NULL}, all the dilutions indicated in a "Sample" variable are used.}

\item{smpdil}{dilution used for the samples.}

\item{antigen}{character string.}

\item{yvar}{character strings for the variables used to fit a standard
curve.}
}
\value{
A list with standards for fitting, background values, sample values,
  indices for samples, and a data frame - for a specified antigen.
}
\description{
Separates specified serial dilutions for fitting, background, and samples
}
\details{
details
}
