% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSmp.R
\name{normalizeSmp}
\alias{normalizeSmp}
\title{Normalize Samples}
\usage{
normalizeSmp(dfout, antigen, fname, pdate, yvar, FUNinv, par, bounds,
  fitflag, trim.flat = TRUE)
}
\arguments{
\item{dfout}{data frame containing samples}

\item{antigen}{character string.}

\item{fname}{name of the file that contains raw data.}

\item{pdate}{date of the processing}

\item{yvar}{character strings for the variables used to fit a standard
curve.}

\item{FUNinv}{inverse function to infer sample concentration}

\item{par}{values of model function parameters}

\item{bounds}{named vector with values for extrema and bounds indicating
"flat" regions of the curve}

\item{fitflag}{flag for the fit as returned by \code{fitStd}}

\item{trim.flat}{logical value determining how the values of \code{yvar} are
trimmed. If \code{TRUE}, they will be trimmed at the bounds where the curve
starts to flatten out (automatically determined as maxima of the third
derivative of the function). If \code{FALSE}, \code{yvar} will be trimmed
at extrema, defined as the range of standards or asymptotes of the fit
(whichever are less extreme).}
}
\value{
a data frame
}
\description{
Calculate sample concentration based on the fit of standard dilutions.
}
\details{

}
